remove update command
Because on Nexus6P content update is not work (return normaly but values are not updated)
Using update is my persistence because it was changed with value only but id increased too (on Nexus5 lolipop).
But on Nexus6P, update command not worked.
Even if used insert command, id was not increased.
So I decided to delete using update command.